+++ config-service/src/main/resources/application.properties
@@ -1,0 +1,3 @@
+spring.application.name=config-server
+server.port=8888
+spring.cloud.config.server.git.uri=https://github.com/cegielskir/en-config-server.git
