--- config-service/src/main/resources/configserver.yml
@@ -1,11 +1,0 @@
-info:
-  component: Config Server
-spring:
-  cloud:
-    config:
-      server:
-        git:
-          uri: https://github.com/cegielskir/en-config-server
-
-server:
-  port: 8888
