--- core-service/src/main/java/com/cgk/engineering/team/coreservice/CoreServiceApplication.java
@@ -1,33 +1,0 @@
-package com.cgk.engineering.team.coreservice;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
-import org.springframework.context.annotation.Bean;
-import springfox.documentation.builders.ApiInfoBuilder;
-import springfox.documentation.builders.PathSelectors;
-import springfox.documentation.builders.RequestHandlerSelectors;
-import springfox.documentation.spi.DocumentationType;
-import springfox.documentation.spring.web.plugins.Docket;
-import springfox.documentation.swagger2.annotations.EnableSwagger2;
-
-@SpringBootApplication
-@EnableDiscoveryClient
-@EnableSwagger2
-public class CoreServiceApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(CoreServiceApplication.class, args);
-	}
-
-	@Bean
-	public Docket swaggerApi() {
-		return new Docket(DocumentationType.SWAGGER_2)
-				.select()
-				.apis(RequestHandlerSelectors.basePackage("com.cgk.engineering.team.coreservice.controller"))
-				.paths(PathSelectors.any())
-				.build()
-				.apiInfo(new ApiInfoBuilder().version("1.0").title("Core API").description("Documentation Employee API v1.0").build());
-	}
-
-}
