--- core-service/src/main/java/com/cgk/engineering/team/coreservice/repository/ArticleRepository.java
@@ -1,10 +1,0 @@
-package com.cgk.engineering.team.coreservice.repository;
-
-import com.cgk.engineering.team.coreservice.model.Article;
-import org.bson.types.ObjectId;
-import org.springframework.data.mongodb.repository.MongoRepository;
-
-public interface ArticleRepository extends MongoRepository<Article, String> {
-    Article findBy_id(ObjectId _id);
-    Article findByTitle(String title);
-}
