--- core-service/src/test/java/com/cgk/engineering/team/coreservice/CoreServiceApplicationTests.java
@@ -1,16 +1,0 @@
-package com.cgk.engineering.team.coreservice;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
-
-@RunWith(SpringRunner.class)
-@SpringBootTest
-public class CoreServiceApplicationTests {
-
-	@Test
-	public void contextLoads() {
-	}
-
-}
