+++ gateway-service/src/main/java/com/cgk/engineering/team/gatewayservice/GatewayServiceApplication.java
@@ -3,10 +3,12 @@ package com.cgk.engineering.team.gatewayservice;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
+import org.springframework.cloud.netflix.zuul.EnableZuulProxy;
 import springfox.documentation.swagger2.annotations.EnableSwagger2;
 
 @SpringBootApplication
 @EnableDiscoveryClient
+@EnableZuulProxy
 @EnableSwagger2
 public class GatewayServiceApplication {
