+++ gateway-service/src/main/resources/bootstrap.yml
@@ -1,6 +1,6 @@
 spring:
   application:
-    name: discovery-service
+    name: gateway-service
   cloud:
     config:
       uri: http://localhost:8888
