+++ main-service/src/test/java/com/cgk/engineering/team/mainservice/MainServiceApplicationTests.java
@@ -1,0 +1,16 @@
+package com.cgk.engineering.team.mainservice;
+
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.junit4.SpringRunner;
+
+@RunWith(SpringRunner.class)
+@SpringBootTest
+public class MainServiceApplicationTests {
+
+	@Test
+	public void contextLoads() {
+	}
+
+}
